Move NodeInserter code out of Contents
As it should have been originally
